Add star-check icon to Explore page
