+++ data/applications/pachyderm.yaml
@@ -10,6 +10,6 @@ spec:
   route: ''
   img: images/pachyderm.svg
   docsLink: https://docs.pachyderm.com/
-  category: Red Hat managed
+  category: Self managed
   support: third party support
   quickStart: ''
