+++ frontend/src/components/OdhCard.scss
@@ -97,6 +97,11 @@
     justify-content: space-between;
   }
 
+  &__title-supported {
+    display: flex;
+    align-items: center;
+  }
+
   &__provider {
     color: var(--pf-global--Color--200);
     font-size: var(--pf-global--FontSize--sm);
