+++ frontend/src/pages/enabledApplications/EnabledApplications.scss
@@ -4,11 +4,6 @@
       .pf-c-card__header {
         padding-right: 0;
       }
-
-      .pf-c-card__title {
-        display: flex;
-        align-items: center;
-      }
     }
   }
 }
